# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API RESTful Laravel Documentation",
        "description": "Project Restaurants API Laravel 12 Documentation",
        "contact": {
            "email": "egarciayavi@hotmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "summary": "Create a new user",
                "description": "Let create a new user with the ",
                "operationId": "92af7d5f95c73e20136cf817e47bde35",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Restaurant created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Login with a user",
                "description": "Let login with a user and return auth token with the ",
                "operationId": "9c7361a02ab87c6f7609f2293c9abf69",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User login successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/get-user": {
            "get": {
                "summary": "Get all users",
                "description": "Return a list of users.",
                "operationId": "ec7eef47785f73e7dbde73d4ec17c7ca",
                "responses": {
                    "200": {
                        "description": "Users List.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "summary": "Logout with a user",
                "description": "Let logout with a user and revoke tokens with the ",
                "operationId": "a983032f3c43ee263567d1b60bd3850d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User logout successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/restaurants": {
            "get": {
                "summary": "Get all restaurants",
                "description": "Return a list of restaurants.",
                "operationId": "0ef65917394f8eb77311d1d085ff31e4",
                "responses": {
                    "200": {
                        "description": "Restaurants List.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Restaurant"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Create a new restaurant",
                "description": "Let create a new restaurant with the ",
                "operationId": "938db3c3288b69d111a27774178532ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Restaurant"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Restaurant created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Restaurant"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/restaurants/{id}": {
            "get": {
                "summary": "To get an specific restaurant",
                "description": "Return a specific restaurant using its ID",
                "operationId": "26e68feb8cf848945261ce292866ec2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the restaurant",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Restaurant"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Restaurant not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Update a restaurant",
                "description": "Update a restaurant using its ID",
                "operationId": "ab7fcec28e65e4b5fda3035e61228114",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the restaurant",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Restaurant"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restaurant updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Restaurant"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Restaurant not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete a restaurant",
                "description": "Delete a restaurant using its ID from the database",
                "operationId": "7270982c262ae0c4e547d5eec368e436",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Restaurant ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "404": {
                        "description": "Restaurant not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Restaurant": {
                "title": "Restaurant",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Restaurant 1"
                    },
                    "address": {
                        "type": "string",
                        "example": "C/Sant Pere d'abanto, 13, 08014 (Barcelona)"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+34601436955"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "elmer"
                    },
                    "email": {
                        "type": "string",
                        "example": "egarciayavi@hotmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password@123"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token Bearer en el formato: 'Bearer {token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}